refactor move out all form field to one file, move out form types to one file
