+++ src/types/authTypes.ts
@@ -27,30 +27,9 @@ export type SignInFormInput = {
   password: string;
 };
 
-export type RegExpPatterns = {
-  [key: string]: string;
-};
-
 export type AuthInputFeldProps = {
   id: string;
   type: string;
   placeholder: string;
   autoComplete: string;
-};
-
-export type MyRegisterOptions = {
-  required: string,
-  minLength?: {
-    value: number,
-    message: string,
-  },
-  maxLength?: {
-    value: number,
-    message: string,
-  },
-  pattern?: {
-    value: string,
-    message: string,
-  },
-
 };
